requirements.txt
python-telegram-bot==20.3
telethon==1.29.1
python-dotenv==1.0.0
